requirements.txt
flask
requests
psycopg2-binary
opentelemetry-distro
opentelemetry-exporter-otlp
